+++ src/app/np-tags-demo/np-tags-demo.component.ts
@@ -100,4 +100,7 @@ export class NpTagsDemoComponent implements OnInit {
     alert(JSON.stringify($event));
   }
 
+  onCreateNewTag($event) {
+    console.log(JSON.stringify($event));
+  }
 }