+++ projects/np-ui-lib/CHANGELOG.md
@@ -5,6 +5,7 @@
 
 # 11.1.2
 - Resolved css issue for vertical line in vertical steps
+- Add onCreateNewTag method in tags component
 
 # 11.1.1
 - Resolved dialog component issue for npAutofocus